requirements.txt
streamlit==1.34.0
st-gsheets-connection==0.0.4
# google-api-python==client-2.129.0
streamlit-authenticator==0.3.2
PyYAML==6.0.1
google-auth==2.29.0
dlt==0.4.11
duckdb==0.9.2
google-auth-httplib2==0.2.0
dbt==duckdb-1.8.0
